- add withTheme hook for the AboutCard - add theme also to native provider
