+++ packages/app/src/Providers.native.tsx
@@ -1,10 +1,12 @@
-import { PropsWithChildren, ReactNode } from 'react';
+// import { PropsWithChildren, ReactNode } from 'react';
 import { NativeBaseProvider } from 'native-base';
 import { ethers } from 'ethers';
 import { Web3Provider } from '@gooddollar/web3sdk-v2';
 import { WalletConnectModal, useWalletConnectModal } from '@walletconnect/modal-react-native';
 // usedapp fix for native
 import LocalStorage from '@usedapp/core/dist/cjs/src/helpers/LocalStorage';
+
+import { nbTheme } from './theme/theme';
 
 type Props = { children: any };
 
@@ -37,7 +39,7 @@ const Web3ProviderWrapper = ({ children }: Props) => {
 };
 export const Providers = ({ children }: Props) => {
   return (
-    <NativeBaseProvider>
+    <NativeBaseProvider theme={nbTheme}>
       <WalletConnectModal projectId={projectId} providerMetadata={providerMetadata} />
       <Web3ProviderWrapper>{children}</Web3ProviderWrapper>
     </NativeBaseProvider>
