+++ packages/app/src/components/theme.ts
@@ -1,0 +1,1 @@
+export { theme as AboutCard } from './AboutCard';
